requirements.txt
python-telegram-bot==13.7
psutil==5.8.0
cryptography==3.4.7
